refactor(Styled Components) - Creating a general button component and using pattern
