+++ src/components/button/pattern.ts
@@ -1,0 +1,3 @@
+export type ButtonProps = {
+    variant?: 'primary' | 'secondary';
+};
