--- src/pages/interviewStart/components/btnIntertviewStartButtonLater/styles.ts
@@ -1,22 +1,0 @@
-import styled from "styled-components";
-import { ButtonYes } from "../btnInterviewStartYes/styles.ts";
-
-// const primaryColor = '#3BB29D';
-
-export const ButtonLater = styled(ButtonYes)`
-    color: 'red'
-`;
-// export const ButtonLater = styled.button`
-//     background-color: '#000';
-//     color: '#000';
-//     border-color: '#fff';
-//     padding: 2.5% 5%;
-//     width: 100%;
-//     border: none;
-//     border-radius: 8px;
-//     color: white;
-//     margin-left: -10%;
-//     font-size: 14px;
-//     margin-top: 10%;
-//     cursor: pointer;
-// `;
