--- src/pages/interviewStart/components/btnInterviewStartYes/styles.ts
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-const primaryColor = '#3BB29D';
-
-export const ButtonYes = styled.button`
-    background-color: ${primaryColor};
-    padding: 2.5% 5%;
-    width: 100%;
-    border: none;
-    border-radius: 8px;
-    color: '#fff';
-    /* margin-left: -10%; */
-    font-size: 14px;
-    cursor: pointer;
-`;
