+++ src/pages/interviewStart/index.jsx
@@ -1,17 +1,7 @@
 import React from "react";
 
-import BtnInterviewStartYes from "./components/btnInterviewStartYes/index.jsx";
-import BtnInterviewStartLater from "./components/btnIntertviewStartButtonLater/index.jsx";
-
-import StyledButton from "./components/styledButton/index.jsx";
-
 import {FilterStartImage, InterviewStart, InterviewStartImage, InterviewStartWelcome, InterviewStartQuestion, InterviewStartText} from "./styles.ts";
-
-const color = { 
-    primary: "#3BB29D",
-    white: "#fff",
-    black: "#000"
-}
+import StyledButton from "../../components/button/index.jsx";
 
 const btnYes = {
     text: "Sim",
@@ -38,11 +28,9 @@ function IntertviewStart(){
                 <InterviewStartText>
                     O mapeamento é importante para conhecermos melhor você, assim escolhendo a viagem perfeita !
                 </InterviewStartText>
-                {/* <BtnInterviewStartYes/>
-                <BtnInterviewStartLater/> */}
-                <StyledButton color={color.white} backgroundColor={color.primary} borderColor={color.primary} link={btnYes.link} text={btnYes.text}></StyledButton>
+                <StyledButton variant="primary" link={btnYes.link} text={btnYes.text}/>
                 <br/>
-                <StyledButton color={color.primary} backgroundColor={color.white} borderColor={color.primary} link={btnLater.link} text={btnLater.text}></StyledButton>
+                <StyledButton variant="secondary" link={btnLater.link} text={btnLater.text}/>
             </InterviewStartWelcome>
         </InterviewStart>
     )
